requirements.txt
nonebot2[fastapi]>=2.4.0
nonebot-adapter-onebot>=2.4.6
httpx>=0.27.0
python-dotenv>=1.0.1
